added fxns into a single file
added fxns into a single file and changed imports in test file
